fix gomod, missing checkapi, moved pkg references
