+++ exporter/splunkhecexporter/config_test.go
@@ -8,6 +8,8 @@ import (
 	"testing"
 	"time"
 
+	translator "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/splunk"
+
 	"github.com/cenkalti/backoff/v5"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -21,7 +23,6 @@ import (
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter/internal/metadata"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/splunk"
 )
 
 func TestLoadConfig(t *testing.T) {
@@ -99,13 +100,13 @@ func TestLoadConfig(t *testing.T) {
 						Sizer:        exporterhelper.RequestSizerTypeItems,
 					}),
 				}),
-				OtelAttrsToHec: splunk.HecToOtelAttrs{
+				OtelAttrsToHec: translator.HecToOtelAttrs{
 					Source:     "mysource",
 					SourceType: "mysourcetype",
 					Index:      "myindex",
 					Host:       "myhost",
 				},
-				HecFields: OtelToHecFields{
+				HecFields: translator.OtelToHecFields{
 					SeverityText:   "myseverityfield",
 					SeverityNumber: "myseveritynumfield",
 				},
