+++ pkg/translator/splunk/common_test.go
